start do services and end with sample of models and DTO

+++ src/main/java/ru/web/laba_web2/repositories/OfferRepository.java
@@ -3,5 +3,5 @@ package ru.web.laba_web2.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.web.laba_web2.models.Offer;
 
-public interface OfferRepository extends JpaRepository<Offer, Long> {
+public interface OfferRepository extends JpaRepository<Offer, Integer> {
 }
