+++ src/main/java/ru/web/laba_web2/repositories/RolesRepository.java
@@ -3,5 +3,5 @@ package ru.web.laba_web2.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.web.laba_web2.models.Roles;
 
-public interface RolesRepository extends JpaRepository<Roles, Long> {
+public interface RolesRepository extends JpaRepository<Roles, Integer> {
 }
